pyproject.toml
[project]
name = "hospital-room-dectection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gymnasium>=0.28.1",
    "joblib>=1.4.2",
    "loguru>=0.7.3",
    "matplotlib>=3.9.2",
    "napari>=0.5.5",
    "omegaconf>=2.3.0",
    "opencv-python>=4.10.0.84",
    "pandas>=2.2.3",
    "plotly>=6.0.0",
    "pydantic>=2.9.2",
    "pyflakes>=3.4.0",
    "pyyaml>=6.0.2",
    "ruff>=0.13.0",
    "scikit-learn>=1.5.1",
    "seaborn>=0.13.2",
    "sentence-transformers>=4.1.0",
    "snakeviz>=2.2.2",
    "tianshou==1.2.0",
    "torch>=2.7.0",
    "torch-geometric>=2.6.1",
    "typer>=0.15.1",
]

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[tool.uv.sources]
torchvision = [
    { index = "pytorch", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torch = [
    { index = "pytorch", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchaudio = [
    { index = "pytorch", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "B",
    "C4",
    "UP",
    "SIM",
    "N",
]

ignore = [
    "E501", # Line too long
    "N805"
]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
docstring-code-format = true
